chore: extract database migration into separate package
