+++ packages/graphback/src/index.ts
@@ -4,15 +4,12 @@
 export * from './GraphQLBackendCreator'
 export * from './IGraphQLBackend'
 
-// DDL layer
-export * from './database/migrations/DataResourcesManager'
-
 // Logger management
-export * from './utils/logger'
-export * from './database';
+// export * from './utils/logger'
 
 export * from '@graphback/codegen-schema'
 export * from '@graphback/codegen-client'
 export * from '@graphback/codegen-core'
 export * from '@graphback/codegen-resolvers'
 export * from '@graphback/runtime'
+export * from '@graphback/db-migrate';
