+++ packages/graphback/src/utils/index.ts
@@ -1,4 +1,2 @@
 
-export * from "./graphqlUtils";
-export * from "./fileUtils";
-export * from "./logger"
+// export * from "./logger"
